Extract the spec normalizer out of the Transformer base class
The load-time normalization methods never touched self — four modules
imported the 978-line Transformer ABC just to call one static method on
it. They now live in spec_normalizer.py as module-level functions, with
the deprecation scan they depend on in spec_scan.py below them.

Dependencies run one way now (validator → spec_normalizer → spec_scan),
so the function-body import that papered over the validator/transformer
cycle is gone. transformer.py drops to 583 lines, validator.py to 1248.

Closes #303
