+++ src/linkml_map/utils/loaders.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import yaml
 
 from linkml_map.datamodel.transformer_model import TransformationSpecification
-from linkml_map.transformer.transformer import Transformer
+from linkml_map.spec_normalizer import normalize_spec
 
 
 def load_specification(path: Path | str) -> TransformationSpecification:
@@ -11,5 +11,5 @@ def load_specification(path: Path | str) -> TransformationSpecification:
         path = str(path)
     with open(path) as f:
         obj = yaml.safe_load(f)
-        Transformer._normalize_spec_dict(obj)
+        normalize_spec(obj)
         return TransformationSpecification(**obj)
